+++ tests/Enums/UserType.php
@@ -5,14 +5,19 @@ namespace BenSampo\Enum\Tests\Enums;
 use BenSampo\Enum\Enum;
 
 /**
+ * @method static static Administrator()
+ * @method static static Moderator()
+ * @method static static Subscriber()
+ * @method static static SuperAdministrator()
+ *
  * @extends Enum<self::*>
  */
 final class UserType extends Enum
 {
-    const Administrator = 0;
-    const Moderator = 1;
-    const Subscriber = 2;
-    const SuperAdministrator = 3;
+    public const Administrator = 0;
+    public const Moderator = 1;
+    public const Subscriber = 2;
+    public const SuperAdministrator = 3;
 
     public function magicInstantiationFromInstanceMethod(): self
     {